🤯 Cognitarium Select query

+++ contracts/okp4-cognitarium/src/msg.rs
@@ -322,6 +322,7 @@ pub struct Results {
 pub enum Value {
     /// # URI
     /// Represents an IRI.
+    #[serde(rename = "uri")]
     URI {
         /// The value of the IRI.
         value: IRI,
@@ -363,7 +364,7 @@ pub struct SelectQuery {
     /// If `None`, there is no limit.
     /// Note: the value of the limit cannot exceed the maximum query limit defined in the store
     /// limitations.
-    pub limit: Option<u64>,
+    pub limit: Option<u32>,
 }
 
 /// # DescribeQuery
